requirements.txt
black==22.6.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
click==8.1.3
cryptography==37.0.4
Flask==2.1.3
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
idna==3.3
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.5.2
pycparser==2.21
PyJWT==2.4.0
PyMySQL==1.0.2
requests==2.28.1
SQLAlchemy==1.4.39
tomli==2.0.1
typing_extensions==4.3.0
urllib3==1.26.10
Werkzeug==2.1.2
zipp==3.8.1
